requirements.txt
# Sabah Informal Housing Detection System
# Python Dependencies

# Web Framework
flask>=2.0.0
flask-cors>=3.0.0

# Geospatial
geopandas>=0.12.0
shapely>=2.0.0
pyproj>=3.4.0
fiona>=1.9.0

# Image Processing
pillow>=9.0.0
opencv-python>=4.7.0
numpy>=1.23.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
kornia>=0.7.0

# Database
supabase>=2.0.0

# Utilities
requests>=2.28.0
tqdm>=4.64.0

# Evaluation (optional)
scikit-learn>=1.2.0
matplotlib>=3.6.0
